Rename to document_to_podcast (#25)
* Rename to document_to_podcast

* Use dash in package name

* Update links
