+++ docs/api.md
@@ -1,7 +1,7 @@
 # API Reference
 
-::: opennotebookllm.preprocessing.data_cleaners
+::: document_to_podcast.preprocessing.data_cleaners
 
-::: opennotebookllm.inference.model_loaders
+::: document_to_podcast.inference.model_loaders
 
-::: opennotebookllm.inference.text_to_text
+::: document_to_podcast.inference.text_to_text
